删除时候存在子菜单的操作

+++ src/domain/OperateTab.php
@@ -224,6 +224,10 @@ class OperateTab
             if (empty($post['id'])) {
                 return ['code' => 0, 'msg' => '参数错误', 'data' => null];
             }
+            $isChild = $this->getDb()->selectOne($this->menuTab, ['parent_id' => $post['id']]);
+            if ($isChild) {
+                return ['code' => 0, 'msg' => '还存在子菜单', 'data' => [$isChild]];
+            }
             $this->getDb()->delete($this->roleAdminTab, ['admin_id' => $post['id']]);
             $result = $this->getDb()->delete($this->adminTab, ['id' => $post['id']]);
             if ($result) {
